+++ src/main/java/com/leanpay/loancalculator/model/response/loan/LoanResponse.java
@@ -5,12 +5,14 @@ import lombok.Data;
 import org.springframework.data.annotation.Id;
 import org.springframework.data.mongodb.core.mapping.Document;
 
+import java.math.BigDecimal;
+
 @Document
 @Data
 @Builder
 public class LoanResponse {
     @Id
     private String id;
-    private double monthlyPayment;
-    private double totalInterestPaid;
+    private BigDecimal monthlyPayment;
+    private BigDecimal totalInterestPaid;
 }
